Split the cli module into submodules, one for each subcommmand
