+++ docs/changes.rst
@@ -2,13 +2,18 @@
 Change Log
 ==========
 
+?.?.?
+-----
+
+- Split the cli module into submodules, one for each subcommmand.
+
 4.0.0
 -----
 
-- Retrieve content from archive, rather than legacy
+- Retrieve content from archive, rather than legacy.
 
 - Make the collectiom version required for 
-  ``neb get [env] [colid] [col_version]``
+  ``neb get [env] [colid] [col_version]``.
   See https://github.com/Connexions/nebuchadnezzar/issues/54
 
 - Warn and prompt if requested not most recently published version.
