+++ nebu/tests/conftest.py
@@ -1,14 +1,37 @@
-# -*- coding: utf-8 -*-
 from pathlib import Path
 
 import pytest
+from click.testing import CliRunner
 
 
 here = Path(__file__).parent
 DATA_DIR = here / 'data'
+CONFIG_FILEPATH = here / 'config.ini'
+
+
+@pytest.fixture(autouse=True)
+def monekypatch_config(monkeypatch):
+    """Point at the testing configuration file"""
+    monkeypatch.setenv('NEB_CONFIG', str(CONFIG_FILEPATH))
 
 
 @pytest.fixture
 def datadir():
     """Returns the path to the data directory"""
     return DATA_DIR
+
+
+@pytest.fixture
+def invoker():
+    """Provides a callable for testing a click enabled function using
+    the click.testing.CliRunner
+
+    """
+    runner = CliRunner()
+    return runner.invoke
+
+
+@pytest.fixture
+def tmpcwd(tmpdir, monkeypatch):
+    monkeypatch.chdir(tmpdir)
+    return Path(str(tmpdir))
