+++ setup.cfg
@@ -1,5 +1,6 @@
 [flake8]
 exclude = *.egg,.state,build
+ignore = F405
 select = E,W,F,N
 
 [aliases]
